Added fewshot data for 10 prompts
Added fewshot data and re-run the evaluation for acquire f version of all figures.
Fixed a bug with the yes preference metric that was causing it to filter out based on the response instead of the outcome.
Fixed a problem with the "proper" label in the zero-shot experiments. Figure 1 is now replaced to account for that.